# server/swagger.yml
openapi: "3.0.3"
info:
  title: SugarBox API by arairon
  version: 0.1.0
  description: A simple storage for SugarCube saves.
  license:
    name: GPL-3.0
    url: https://opensource.org/license/gpl-3-0
  contact:
    name: arai.icu
    url: https://arai.icu
    email: arai@arai.icu
servers:
  - url: https://sugarbox.arai.icu/api
  #- url: http://127.0.0.1:3000/api
paths:
...
  /games:
    get:
      summary: Get all games for the authenticated user
      description: Retrieves all games owned by the authenticated user.
      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
          description: Bearer token in the format "Bearer ACCESSTOKEN"
      responses:
        "200":
          description: Successfully retrieved games
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GameSchema"
        "403":
          description: Invalid authentication data
        "500":
          description: Internal server error
  /games/uuid/{gameId}:
    get:
      summary: Get a game by UUID
      description: Retrieves a specific game by its UUID for the authenticated user.
      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
          description: Bearer token in the format "Bearer ACCESSTOKEN"
        - in: path
          name: gameId
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the game to retrieve.
      responses:
        "200":
          description: Successfully retrieved the game
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameSchema"
        "400":
          description: Invalid UUID
        "404":
          description: Game not found
        "403":
          description: Invalid authentication data
        "500":
          description: Internal server error

    patch:
      summary: Update a game by UUID
      description: Updates a specific game by its UUID for the authenticated user.
      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
          description: Bearer token in the format "Bearer ACCESSTOKEN"
        - in: path
          name: gameId
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the game to update.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameUpdateSchema"
      responses:
        "200":
          description: Successfully updated the game
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameSchema"
        "400":
          description: Invalid UUID or validation error
        "403":
          description: Invalid authentication data
        "500":
          description: Internal server error

    delete:
      summary: Delete a game by UUID
      description: Deletes a specific game by its UUID for the authenticated user.
      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
          description: Bearer token in the format "Bearer ACCESSTOKEN"
        - in: path
          name: gameId
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the game to delete.
      responses:
        "200":
          description: Successfully deleted the game
        "400":
          description: Invalid UUID
        "403":
          description: Invalid authentication data
        "500":
          description: Internal server error
  /games/new:
    post:
      summary: Create a new game
      description: Creates a new game for the authenticated user.
      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
          description: Bearer token in the format "Bearer ACCESSTOKEN"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GameCreateSchema"
      responses:
        "200":
          description: Successfully created a new game
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameSchema"
        "400":
          description: Validation error
        "403":
          description: Invalid authentication data
        "500":
          description: Internal server error
  /chars:
    get:
      summary: Get all characters for the authenticated user
      description: Retrieves all characters owned by the authenticated user.
      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
          description: Bearer token in the format "Bearer ACCESSTOKEN"
      responses:
        "200":
          description: Successfully retrieved characters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CharSchema"
        "403":
          description: Invalid authentication data
        "500":
          description: Internal server error
  /chars/uuid/{charId}:
    get:
      summary: Get a character by UUID
      description: Retrieves a specific character by its UUID for the authenticated user.
      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
          description: Bearer token in the format "Bearer ACCESSTOKEN"
        - in: path
          name: charId
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the character to retrieve.
      responses:
        "200":
          description: Successfully retrieved the character
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CharSchema"
        "400":
          description: Invalid UUID
        "404":
          description: Character not found
        "403":
          description: Invalid authentication data
        "500":
          description: Internal server error

    patch:
      summary: Update a character by UUID
      description: Updates a specific character by its UUID for the authenticated user.
      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
          description: Bearer token in the format "Bearer ACCESSTOKEN"
        - in: path
          name: charId
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the character to update.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CharUpdateSchema"
      responses:
        "200":
          description: Successfully updated the character
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CharSchema"
        "400":
          description: Invalid UUID or validation error
        "403":
          description: Invalid authentication data
        "500":
          description: Internal server error

    delete:
      summary: Delete a character by UUID
      description: Deletes a specific character by its UUID for the authenticated user.
      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
          description: Bearer token in the format "Bearer ACCESSTOKEN"
        - in: path
          name: charId
          required: true
          schema:
            type: string
            format: uuid
          description: The UUID of the character to delete.
      responses:
        "200":
          description: Successfully deleted the character
        "400":
          description: Invalid UUID
        "403":
          description: Invalid authentication data
        "500":
          description: Internal server error
  /chars/new:
    post:
      summary: Create a new character
      description: Creates a new character for the authenticated user.
      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
          description: Bearer token in the format "Bearer ACCESSTOKEN"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CharCreateSchema"
      responses:
        "200":
          description: Successfully created a new character
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CharSchema"
        "400":
          description: Validation error
        "403":
          description: Invalid authentication data
        "500":
          description: Internal server error
...
components:
  schemas:
...
    GameSchema:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the game.
        uuid:
          type: string
          format: uuid
          description: UUID of the game.
        name:
          type: string
          maxLength: 100
          description: Name of the game.
        shortname:
          type: string
          maxLength: 100
          description: Short name of the game.
        description:
          type: string
          maxLength: 256
          description: Description of the game.
        ownerId:
          type: integer
          description: ID of the owner of the game.
        paths:
          type: string
          description: JSON string representing an array of game paths.
        archived:
          type: boolean
          description: Whether the game is archived.
        archivedAt:
          type: string
          format: date-time
          description: Timestamp when the game was archived.
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the game was last updated.
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the game was created.
    GameCreateSchema:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
          description: UUID of the game.
        name:
          type: string
          maxLength: 100
          description: Name of the game.
        shortname:
          type: string
          maxLength: 100
          description: Short name of the game.
        description:
          type: string
          maxLength: 256
          description: Description of the game.
        paths:
          type: string
          description: JSON string representing an array of game paths.
    GameUpdateSchema:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
          description: Name of the game.
        shortname:
          type: string
          maxLength: 100
          description: Short name of the game.
        description:
          type: string
          maxLength: 256
          description: Description of the game.
        paths:
          type: string
          description: JSON string representing an array of game paths.
    CharSchema:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the character.
        uuid:
          type: string
          format: uuid
          description: UUID of the character.
        name:
          type: string
          maxLength: 100
          description: Name of the character.
        ownerId:
          type: integer
          description: ID of the owner of the character.
        gameId:
          type: string
          format: uuid
          description: UUID of the game associated with the character.
        slots:
          type: string
          description: JSON string representing an array of UUIDs or empty strings.
        archived:
          type: boolean
          description: Whether the character is archived.
        archivedAt:
          type: string
          format: date-time
          description: Timestamp when the character was archived.
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the character was last updated.
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the character was created.
    CharCreateSchema:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
          description: UUID of the character.
        name:
          type: string
          maxLength: 100
          description: Name of the character.
        gameId:
          type: string
          format: uuid
          description: UUID of the game associated with the character.
        slots:
          type: string
          description: JSON string representing an array of UUIDs or empty strings.
    CharUpdateSchema:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
          description: Name of the character.
        gameId:
          type: string
          format: uuid
          description: UUID of the game associated with the character.
        slots:
          type: string
          description: JSON string representing an array of UUIDs or empty strings.
        archived:
          type: boolean
          description: Whether the character is archived.
...